# Pertemuan 4/project_mysql/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TokoBersama API",
    "version": "1.0.0",
    "description": "REST API untuk platform marketplace TokoBersama. Mengelola kategori produk, seller, produk, dan order."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "components": {
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": {}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "data": { "nullable": true, "example": null }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "id_produk": { "type": "integer", "example": 1 },
          "jumlah": { "type": "integer", "example": 2 },
          "total_harga": { "type": "number", "example": 9998000 },
          "status": {
            "type": "string",
            "enum": ["pending", "proses", "selesai", "batal"],
            "example": "pending"
          },
          "nama_produk": { "type": "string", "example": "Samsung Galaxy A55" },
          "harga_satuan": { "type": "number", "example": 4999000 },
          "nama_kategori": { "type": "string", "example": "Elektronik" },
          "nama_toko": { "type": "string", "example": "TechZone Official" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/order": {
      "get": {
        "tags": ["Order"],
        "summary": "Ambil semua order beserta detail produk",
        "responses": {
          "200": {
            "description": "Daftar semua order",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Order" }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Order"],
        "summary": "Buat order baru",
        "description": "Menggunakan **MySQL Transaction**. Otomatis mengurangi stok produk dan menghitung `total_harga = harga × jumlah`. Menggunakan `SELECT ... FOR UPDATE` untuk mencegah race condition.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id_produk", "jumlah"],
                "properties": {
                  "id_produk": { "type": "integer", "example": 1 },
                  "jumlah": { "type": "integer", "minimum": 1, "example": 2 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Order" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Stok tidak mencukupi / validasi gagal",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Produk tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/order/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "example": 1
        }
      ],
      "get": {
        "tags": ["Order"],
        "summary": "Ambil detail order berdasarkan ID",
        "responses": {
          "200": {
            "description": "Detail order",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "properties": {
                        "data": { "$ref": "#/components/schemas/Order" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Order tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Order"],
        "summary": "Hapus/batalkan order",
        "description": "Jika status order masih `pending`, stok produk akan dikembalikan secara otomatis.",
        "responses": {
          "200": {
            "description": "Order berhasil dihapus",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "404": {
            "description": "Order tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/order/{id}/status": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "example": 1
        }
      ],
      "put": {
        "tags": ["Order"],
        "summary": "Update status order",
        "description": "Alur status yang valid: `pending` → `proses` → `selesai` atau `batal`. Order yang sudah `selesai` atau `batal` tidak bisa diubah lagi.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["pending", "proses", "selesai", "batal"],
                    "example": "proses"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status berhasil diupdate",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": {
            "description": "Transisi status tidak valid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Order tidak ditemukan",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}